frontend revamps
